rename: api에서 전달된 정렬 조건을 필드 명으로 변경하는 기능 네이밍 변경
