+++ backend/ddang/src/main/java/com/ddang/ddang/auction/configuration/exception/UnsupportedSortParameterException.java
@@ -1,0 +1,8 @@
+package com.ddang.ddang.auction.configuration.exception;
+
+public class UnsupportedSortParameterException extends IllegalArgumentException {
+
+    public UnsupportedSortParameterException(final String message) {
+        super(message);
+    }
+}
