--- backend/ddang/src/main/java/com/ddang/ddang/auction/configuration/exception/UnsupportedSortTypeException.java
@@ -1,8 +1,0 @@
-package com.ddang.ddang.auction.configuration.exception;
-
-public class UnsupportedSortTypeException extends IllegalArgumentException {
-
-    public UnsupportedSortTypeException(final String message) {
-        super(message);
-    }
-}
